+++ src/tiled/mapdocument.h
@@ -32,6 +32,7 @@
 
 #include <QList>
 #include <QRegion>
+#include <QSet>
 
 #include <memory>
 
@@ -257,6 +258,9 @@ public:
     bool templateAllowed(const ObjectTemplate *objectTemplate) const;
 
     void checkIssues() override;
+
+    QSet<int> expandedGroupLayers;
+    QSet<int> expandedObjectLayers;
 
 signals:
     /**