+++ src/tiled/regionvaluetype.cpp
@@ -59,6 +59,7 @@ QVector<QRect> RegionValueType::rects() const
     rects.reserve(static_cast<int>(mRegion.end() - mRegion.begin()));
     for (const QRect &rect : mRegion)
         rects.append(rect);
+    return rects;
 #else
     return QVector<QRect>(mRegion.begin(), mRegion.end());
 #endif